Add files via upload
Now reads in two neighbouring maps, so as to properly include regions that straddle the map boundary. Such regions are included if their central longitude lies within [0, 360) of the original map, so they are only included once in the full simulation. The output has a different format from before, with separate Br and mask arrays for each region (the mask array is now binary). I also added an output of the sign of the leading (westward) polarity for each region.